Server folders, drag-and-drop reordering, and UI bug fixes
Add server folder system with drag-into-folder, reorder, edit/delete (localStorage-persisted)
Fix channel name showing "unnamed" for non-JSON metadata (raw base64 fallback)
Fix category drag-and-drop with pointer-based collision detection
Fix server icon disappearing on reload (CSS.Translate vs CSS.Transform)
Fix invite modal flicker on dropdown change
Fix delete emoji confirm dialog hidden behind settings overlay (z-index)
Bump member list fetch limit to 100

+++ src/models.rs
@@ -786,7 +786,7 @@ pub struct Invite {
 #[derive(Debug, Deserialize)]
 pub struct CreateInviteRequest {
     pub max_uses: Option<i32>,
-    pub expires_in_hours: Option<i64>,
+    pub expires_in_hours: Option<f64>,
 }
 
 #[derive(Debug, Serialize)]
